+++ HISTORY.md
@@ -3,9 +3,12 @@ History
 
 ## 0.0.2
 
-* Extension of API with: `$css`, `$visible`, `$hidden`.
+* Extension of API with: `$css`, `$visible`, `$hidden`.  [*[atimb][]*]
 
 ## 0.0.1
 
 * Initial release with API: `$val`, `$class`, `$attr`, `$prop`, `$html`,
-  `$text`.
+  `$text`. [*[ryan-roemer][]*]
+
+[atimb]: https://github.com/atimb
+[ryan-roemer]: https://github.com/ryan-roemer