+++ HISTORY.md
@@ -3,6 +3,8 @@ History
 
 ## 0.0.3
 
+* Switch to manual plugin injection for AMD (matches Node.js method).
+  [*[@ryan-roemer][]*]
 * Add fallback to "style" property for `$css`. [*[@ryan-roemer][]*]
 * Add `contains` support for `$css` assert. [*[@ryan-roemer][]*]
 * Add support for Node.js with JsDom. [*[@ryan-roemer][]*]